refactor(logging): move buildLogger to internal/logging so the desktop can use it
